+++ Assets/Neoxider/Scripts/Audio/AudioSimple/PlayAudioBtn.cs
@@ -1,6 +1,6 @@
 using Neo.Extensions;
 using UnityEngine;
-using UnityEngine.UI;
+using UnityEngine.EventSystems;
 
 namespace Neo
 {
@@ -10,34 +10,31 @@ namespace Neo
         [NeoDoc("Audio/PlayAudioBtn.md")]
         [CreateFromMenu("Neoxider/Audio/PlayAudioBtn")]
         [AddComponentMenu("Neoxider/" + "Audio/" + nameof(PlayAudioBtn))]
-        public class PlayAudioBtn : MonoBehaviour
+        public class PlayAudioBtn : MonoBehaviour, IPointerEnterHandler, IPointerExitHandler, IPointerDownHandler,
+            IPointerUpHandler, IPointerClickHandler, ISelectHandler, IDeselectHandler, ISubmitHandler
         {
+            public enum TriggerMode
+            {
+                PointerClick,
+                PointerEnter,
+                PointerExit,
+                PointerDown,
+                PointerUp,
+                Select,
+                Deselect,
+                Submit,
+                Manual
+            }
+
             [Header("Legacy Mode (by ID)")] [SerializeField]
             private int _idClip;
 
             [Header("New Mode (by Clip)")] [SerializeField]
             private AudioClip[] _clips;
 
+            [Header("Trigger")] [SerializeField] private TriggerMode _triggerMode = TriggerMode.PointerClick;
             [SerializeField] private bool _useRandomClip;
             [SerializeField] private float _volume = 1f;
-
-            [SerializeField] [GetComponent] private Button _button;
-
-            private void OnEnable()
-            {
-                if (_button != null)
-                {
-                    _button.onClick.AddListener(AudioPlay);
-                }
-            }
-
-            private void OnDisable()
-            {
-                if (_button != null)
-                {
-                    _button.onClick.RemoveListener(AudioPlay);
-                }
-            }
 
             /// <summary>Plays the sound. If clips are set and useRandomClip is true, picks random; otherwise uses legacy ID mode.</summary>
             public void AudioPlay()
@@ -67,6 +64,61 @@ namespace Neo
                 {
                     AM.I?.Play(_idClip);
                 }
+            }
+
+            public void OnPointerEnter(PointerEventData eventData)
+            {
+                TryTrigger(TriggerMode.PointerEnter);
+            }
+
+            public void OnPointerExit(PointerEventData eventData)
+            {
+                TryTrigger(TriggerMode.PointerExit);
+            }
+
+            public void OnPointerDown(PointerEventData eventData)
+            {
+                TryTrigger(TriggerMode.PointerDown);
+            }
+
+            public void OnPointerUp(PointerEventData eventData)
+            {
+                TryTrigger(TriggerMode.PointerUp);
+            }
+
+            public void OnPointerClick(PointerEventData eventData)
+            {
+                if (eventData != null && eventData.button != PointerEventData.InputButton.Left)
+                {
+                    return;
+                }
+
+                TryTrigger(TriggerMode.PointerClick);
+            }
+
+            public void OnSelect(BaseEventData eventData)
+            {
+                TryTrigger(TriggerMode.Select);
+            }
+
+            public void OnDeselect(BaseEventData eventData)
+            {
+                TryTrigger(TriggerMode.Deselect);
+            }
+
+            public void OnSubmit(BaseEventData eventData)
+            {
+                TryTrigger(TriggerMode.Submit);
+            }
+
+            private void TryTrigger(TriggerMode triggerMode)
+            {
+                if (_triggerMode != triggerMode)
+                {
+                    return;
+                }
+
+                AudioPlay();
             }
         }
     }